fix(gateway): harden spend limit persistence
